기능개발 : 사용자 메뉴 추가 및 메뉴 클릭 시 컨텐츠 렌더링 처리
